kernel: clear all compiler warnings (31 on x86, 4 on x86_64 -> 0)
- struct task: align pending_signals to 8 bytes. All 24 i686
  -Watomic-alignment warnings were 8-byte __atomic ops on this u64 field,
  which is only 4-byte aligned in the struct on 32-bit (the lock-prefixed op
  is correct but warns and pays a cache penalty). Tasks are kmalloc'd in
  chunks, so the small sizeof bump is harmless; no asm hardcodes task field
  offsets after this field.
- ext4_commit_inode: cast size to u64 before >>32 (shift-count-overflow on
  32-bit where usize is 32 bits; high bits are always 0 there anyway).
- mc readdir: compare the readdir return against (usize)-1, not (u64)-1
  (always-false on 32-bit, so the error case was never detected).
- remove dead code: tcp_queue_retransmit (never called; retransmit is driven
  off the send buffer), journal replay `count`, usb_xhci `cc`/`length`.
- m36_diag: guard decode_le64 under __x86_64__ (only the 64-bit GDB register
  check uses it; i686 decodes 4-byte regs inline).

Both arches build warning-free and stay green (x86 742/0, x86_64 743/0).

+++ kernel/fs/ext4.c
@@ -1002,7 +1002,7 @@ static int ext4_vfs_setattr(struct vfs_node *node) {
   inode.i_size = (u32)node->inode->size;
   inode.i_atime = node->inode->atime; inode.i_mtime = node->inode->mtime; inode.i_ctime = node->inode->ctime;
   if (fs->features_ro_compat & EXT4_FEATURE_RO_COMPAT_HUGE_FILE) {
-    inode.i_dir_acl = (u32)(node->inode->size >> 32);
+    inode.i_dir_acl = (u32)((u64)node->inode->size >> 32);
   }
   return ext4_write_inode(fs, info->inode_num, &inode);
 }
